Donation Bot/requirements.txt
aiogram==3.4.1
httpx==1.0.0b0
python-dotenv==1.0.1
